AssimpView:  - postprocessing configuration is no (partially) configurable.  - adding linsk to Assimp homepage/bugtracker/forums.  - updating about dialog.  - fixing colors of texture view (tiny array overrun ...).  - minor UI theme adjustments.  - 'setup file extensions' works again.  - normal smooth limit is now used for primary import, too.  - minor spelling fixes in the UI.
git-svn-id: https://assimp.svn.sourceforge.net/svnroot/assimp/trunk@417 67173fc5-114c-0410-ac8e-9d2fd5bffc1f

+++ mkutil/mkrev.bat
@@ -15,14 +15,13 @@ set /p addtext= < tmpfile.txt
 del /q tmpfile.txt
 cd ..\mkutil
 
-echo #define SVNRevision > revision.h
+echo #define SVNRevision > tmpfile.txt
 
-if exist tmpfile.txt del /q tmpfile.txt
-for /f "delims=" %%l in (revision.h) Do (
-      echo %%l %addtext% >> tmpfile.txt
+if exist revision.h del /q revision.h
+for /f "delims=" %%l in (tmpfile.txt) Do (
+      echo %%l %addtext% >> revision.h
 )
-del /q revision.h
-ren tmpfile.txt revision.h
+del /q tmpfile.txt
